add release applications and docker
