+++ release/run
@@ -7,8 +7,12 @@ use Symfony\Component\Console\Application;
 $application = new Application('release', '1.0.0');
 
 $application->add((new \EMS\Release\Command\ComposerUpdate()));
+
 $application->add((new \EMS\Release\Command\GitHubBranchesCommand()));
+$application->add((new \EMS\Release\Command\GithubReleaseApplications()));
+$application->add((new \EMS\Release\Command\GithubReleaseDocker()));
 $application->add((new \EMS\Release\Command\GithubReleasePackages()));
+
 $application->add((new \EMS\Release\Command\PackagistInfo()));
 
 $application->run();
